Rename shared test package
Having a partially common package name was causing BuildConfig getting defined multiple times
