+++ ground/src/androidTest/java/com/google/android/ground/AcceptTermsOfServiceTest.java
@@ -24,9 +24,9 @@ import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.not;
 
-import com.google.android.ground.test.FakeData;
-import com.google.android.ground.test.persistence.remote.FakeRemoteDataStore;
-import com.google.android.ground.test.system.auth.FakeAuthenticationManager;
+import com.sharedtest.FakeData;
+import com.sharedtest.persistence.remote.FakeRemoteDataStore;
+import com.sharedtest.system.auth.FakeAuthenticationManager;
 import dagger.hilt.android.testing.HiltAndroidTest;
 import javax.inject.Inject;
 import org.junit.Test;
