+++ ground/src/androidTest/java/com/google/android/ground/MapTypeDialogTest.kt
@@ -22,9 +22,9 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.ground.repository.MapsRepository
-import com.google.android.ground.test.FakeData
-import com.google.android.ground.test.system.auth.FakeAuthenticationManager
 import com.google.common.truth.Truth
+import com.sharedtest.FakeData
+import com.sharedtest.system.auth.FakeAuthenticationManager
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Test
 import javax.inject.Inject
