+++ ground/src/androidTest/java/com/google/android/ground/TestRemoteStorageModule.java
@@ -20,9 +20,9 @@ import com.google.android.ground.persistence.remote.RemoteDataStore;
 import com.google.android.ground.persistence.remote.RemoteStorageManager;
 import com.google.android.ground.persistence.remote.RemoteStorageModule;
 import com.google.android.ground.persistence.uuid.OfflineUuidGenerator;
-import com.google.android.ground.test.persistence.remote.FakeRemoteDataStore;
-import com.google.android.ground.test.persistence.remote.FakeRemoteStorageManager;
-import com.google.android.ground.test.persistence.uuid.FakeUuidGenerator;
+import com.sharedtest.persistence.remote.FakeRemoteDataStore;
+import com.sharedtest.persistence.remote.FakeRemoteStorageManager;
+import com.sharedtest.persistence.uuid.FakeUuidGenerator;
 import dagger.Binds;
 import dagger.Module;
 import dagger.hilt.components.SingletonComponent;
