+++ ground/src/test/java/com/google/android/ground/MainViewModelTest.java
@@ -16,9 +16,9 @@
 
 package com.google.android.ground;
 
-import static com.google.android.ground.test.FakeData.TERMS_OF_SERVICE;
-import static com.google.android.ground.test.FakeData.USER;
 import static com.google.common.truth.Truth.assertThat;
+import static com.sharedtest.FakeData.TERMS_OF_SERVICE;
+import static com.sharedtest.FakeData.USER;
 
 import android.content.SharedPreferences;
 import androidx.navigation.NavDirections;
@@ -26,12 +26,12 @@ import com.google.android.ground.repository.TermsOfServiceRepository;
 import com.google.android.ground.repository.UserRepository;
 import com.google.android.ground.system.auth.SignInState;
 import com.google.android.ground.system.auth.SignInState.State;
-import com.google.android.ground.test.TestObservers;
-import com.google.android.ground.test.persistence.remote.FakeRemoteDataStore;
-import com.google.android.ground.test.system.auth.FakeAuthenticationManager;
 import com.google.android.ground.ui.common.Navigator;
 import com.google.android.ground.ui.home.HomeScreenFragmentDirections;
 import com.google.android.ground.ui.signin.SignInFragmentDirections;
+import com.sharedtest.TestObservers;
+import com.sharedtest.persistence.remote.FakeRemoteDataStore;
+import com.sharedtest.system.auth.FakeAuthenticationManager;
 import dagger.hilt.android.testing.HiltAndroidTest;
 import io.reactivex.observers.TestObserver;
 import java.util.NoSuchElementException;
