+++ ground/src/test/java/com/google/android/ground/TestWorkManagerModule.java
@@ -18,7 +18,7 @@ package com.google.android.ground;
 
 import androidx.work.WorkManager;
 import com.google.android.ground.persistence.sync.WorkManagerModule;
-import com.google.android.ground.test.persistence.sync.FakeWorkManager;
+import com.sharedtest.persistence.sync.FakeWorkManager;
 import dagger.Module;
 import dagger.Provides;
 import dagger.hilt.components.SingletonComponent;
