+++ ground/src/test/java/com/google/android/ground/persistence/remote/firestore/schema/SubmissionConverterTest.java
@@ -35,11 +35,11 @@ import com.google.android.ground.model.task.MultipleChoice;
 import com.google.android.ground.model.task.MultipleChoice.Cardinality;
 import com.google.android.ground.model.task.Task;
 import com.google.android.ground.persistence.remote.DataStoreException;
-import com.google.android.ground.test.FakeData;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.firebase.Timestamp;
 import com.google.firebase.firestore.DocumentSnapshot;
+import com.sharedtest.FakeData;
 import java.util.Date;
 import java8.util.Optional;
 import org.junit.Test;
