+++ ground/src/test/java/com/google/android/ground/repository/MapsRepositoryTest.kt
@@ -17,8 +17,8 @@ package com.google.android.ground.repository
 
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.ground.BaseHiltTest
-import com.google.android.ground.test.TestObservers.observeUntilFirstChange
 import com.google.common.truth.Truth.assertThat
+import com.sharedtest.TestObservers.observeUntilFirstChange
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Test
 import org.junit.runner.RunWith
