+++ ground/src/test/java/com/google/android/ground/repository/SurveyRepositoryTest.java
@@ -23,8 +23,8 @@ import com.google.android.ground.BaseHiltTest;
 import com.google.android.ground.model.Survey;
 import com.google.android.ground.persistence.local.LocalDataStore;
 import com.google.android.ground.persistence.local.LocalDataStoreModule;
-import com.google.android.ground.test.persistence.remote.FakeRemoteDataStore;
 import com.google.common.collect.ImmutableMap;
+import com.sharedtest.persistence.remote.FakeRemoteDataStore;
 import dagger.hilt.android.testing.BindValue;
 import dagger.hilt.android.testing.HiltAndroidTest;
 import dagger.hilt.android.testing.UninstallModules;
