+++ ground/src/test/java/com/google/android/ground/repository/UserRepositoryTest.java
@@ -23,8 +23,8 @@ import com.google.android.ground.model.Role;
 import com.google.android.ground.model.Survey;
 import com.google.android.ground.persistence.local.LocalDataStore;
 import com.google.android.ground.persistence.local.LocalValueStore;
-import com.google.android.ground.test.FakeData;
-import com.google.android.ground.test.system.auth.FakeAuthenticationManager;
+import com.sharedtest.FakeData;
+import com.sharedtest.system.auth.FakeAuthenticationManager;
 import dagger.hilt.android.testing.HiltAndroidTest;
 import java.util.NoSuchElementException;
 import javax.inject.Inject;
