+++ sharedTest/src/main/java/com/sharedtest/FakeData.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.test
+package com.sharedtest
 
 import com.google.android.ground.model.AuditInfo
 import com.google.android.ground.model.Survey
