+++ sharedTest/src/main/java/com/sharedtest/LiveDataTestUtil.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.ground.test
+package com.sharedtest
 
 import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.LiveData
