+++ sharedTest/src/main/java/com/sharedtest/TestScheduler.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.ground.test;
+package com.sharedtest;
 
 import androidx.test.espresso.IdlingRegistry;
 import com.google.android.ground.rx.Schedulers;
